Move mosaic selector to separate folder
